Update XenAPI for better Python compatibility
